fix: split Tree assign Widgetbook stories under NCL cap (#4498)
catalog_panels.dart exceeded the 1000 non-comment-line Quality gate after
Tree assign use cases were inlined; extract them to a sibling catalog part.
